mocks and tests added

+++ backend/services/category.py
@@ -1,3 +1,4 @@
+from exeptions.category import CategoryNotFoundError
 from repositories.category import CategoryRepository
 from schemas.category import CategoryCreateSchema, CategorySchema, CategoryUdpateSchema
 from sqlalchemy.orm import Session
@@ -19,10 +20,13 @@ class CategoryService:
         self.db.commit()
         return CategorySchema.model_validate(category_orm)
 
-    def update_categoty(
+    def update_category(
         self, id: str, category_update: CategoryUdpateSchema
     ) -> CategorySchema:
         category_orm = self.category_repository.get_by_id(id=id)
+
+        if category_orm is None:
+            raise CategoryNotFoundError(f"Категория {id} не найдена")
 
         category_orm.name = category_update.name
         self.db.commit()
@@ -30,5 +34,9 @@ class CategoryService:
 
     def delete_category(self, id: str) -> None:
         category_orm = self.category_repository.get_by_id(id)
+
+        if category_orm is None:
+            raise CategoryNotFoundError(f"Категория {id} не найдена")
+
         self.db.delete(category_orm)
         self.db.commit()
